requirements.txt
Django==1.6.2
Fabric==1.8.3
MySQL-python==1.2.5
South==0.8.4
Unidecode==0.04.14
chardet==2.2.1
csvImporter==0.1.3.2
-e git://github.com/numerodix/django-access-log.git@13408d0976f0c2a8aa28b99aa7b2fff7f321c9d4#egg=django_access_log-master
django-adaptors==0.2.5
django-allauth==0.16.1
django-appconf==0.6
django-autocomplete-light==1.4.13
django-autoslug==1.7.2
django-cities-light==2.3.0
django-csvimport==1.0
django-debug-toolbar==1.0.1
django-documents==0.0.3
django-extensions==1.3.3
django-generic-m2m==0.2.2
django-grappelli==2.5.2
django-lockdown==1.0
django-session-security==2.1.6
django-smart-selects==1.0.9
django-suit==0.2.7
ecdsa==0.11
lxml==3.3.3
oauthlib==0.6.1
paramiko==1.12.3
pycrypto==2.6.1
python-dateutil==2.2
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
sqlparse==0.1.11
wsgiref==0.1.2
xlrd==0.9.2
